refactoring: move ai-prompts in one file
